Fixed major bug in logic of compute_moves.c - now seems to be calculating correctly and running on GPU. disappointingly not that fast yet though

+++ compute_moves.h
@@ -4,7 +4,7 @@
 typedef int transform_t;
 const int NUM_TRANSFORMS = 65536;
 
-void compute_moves(uint64_t *allBoards, size_t tree_size ,uint64_t layer_num, transform_t left_move_transforms[NUM_TRANSFORMS], transform_t right_move_transforms[NUM_TRANSFORMS]);
+void compute_moves(uint64_t *allBoards, size_t tree_size ,unsigned int layer_num, transform_t left_move_transforms[NUM_TRANSFORMS], transform_t right_move_transforms[NUM_TRANSFORMS]);
 
 
 #endif
